% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_lhz.R
\name{download_lhz_shapefiles}
\alias{download_lhz_shapefiles}
\title{Download Malawi livelihood zones shapefiles from Humanitarian Data Exchange}
\usage{
download_lhz_shapefiles(.unzip = TRUE)
}
\arguments{
\item{.unzip}{Logical. Should zip file be unzipped and extracted? Default to
TRUE}
}
\value{
If \code{.unzip}, list of shapefiles within the downloaded zip file else
path to downloaded zip file. (Invisible) Shapefiles zip file downloaded to
specified folder.
}
\description{
Download Malawi livelihood zones shapefiles from Humanitarian Data Exchange
}
\examples{
download_lhz_shapefiles()

}
