% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_boundaries.R
\name{get_boundaries_name}
\alias{get_boundaries_name}
\title{Get administrative boundaries layer name}
\usage{
get_boundaries_name(files = download_boundaries_shapefiles(), adm = NULL)
}
\arguments{
\item{files}{List of shapefiles within a downloaded zip file}

\item{adm}{Integer. Administrative level required. Either \code{0} for country,
\code{1} for regions, \code{2} for districts, or \code{3} for traditional authority areas.}
}
\value{
Layer name for specified administrative level boundary
}
\description{
Get administrative boundaries layer name
}
\examples{
get_boundaries_name(adm = 0)

}
