% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_boundaries.R
\name{get_boundary}
\alias{get_boundary}
\title{Retrieve Malawi boundary layer}
\usage{
get_boundary(id, adm = NULL)
}
\arguments{
\item{id}{Character vector for unique identifier of administrative
boundaries data for Malawi from Humanitarian Data Exchange}

\item{adm}{Integer. Administrative level required. Either \code{0} for country,
\code{1} for regions, \code{2} for districts, or \code{3} for traditional authority areas.}
}
\value{
An sf object for specific boundary of Malawi
}
\description{
Retrieve Malawi boundary layer
}
\examples{
get_boundary(id = "cod-ab-mwi", adm = 0)

}
