% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_boundaries.R
\name{get_country}
\alias{get_country}
\title{Retrieve Malawi country boundaries}
\usage{
get_country(layer = get_boundaries_name(adm = 0))
}
\arguments{
\item{layer}{Name of layer for country boundaries}
}
\value{
An sf object for the country borders of Malawi
}
\description{
Retrieve Malawi country boundaries
}
\examples{
get_country()

}
