% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_boundaries.R
\name{get_districts}
\alias{get_districts}
\title{Retrieve Malawi districts boundaries}
\usage{
get_districts(id)
}
\arguments{
\item{id}{Character vector for unique identifier of administrative
boundaries data for Malawi from Humanitarian Data Exchange}
}
\value{
An sf object for the districts borders of Malawi
}
\description{
Retrieve Malawi districts boundaries
}
\examples{
get_districts(id = "cod-ab-mwi")

}
