% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_lhz.R
\name{get_lhz}
\alias{get_lhz}
\title{Retrieve Malawi livelihood zones}
\usage{
get_lhz(layer = get_lhz_name())
}
\arguments{
\item{layer}{Name of layer for livelihood zones}
}
\value{
An sf object for the livelihood zones of Malawi
}
\description{
Retrieve Malawi livelihood zones
}
\examples{
get_lhz()

}
