% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_lhz.R
\name{get_lhz_name}
\alias{get_lhz_name}
\title{Get livelihoods zones layer name}
\usage{
get_lhz_name(files = download_lhz_shapefiles())
}
\arguments{
\item{files}{List of shapefiles within a downloaded zip file}
}
\value{
Layer name for livelihood zones shapefile
}
\description{
Get livelihoods zones layer name
}
\examples{
get_lhz_name()

}
