% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_boundaries.R
\name{get_regions}
\alias{get_regions}
\title{Retrieve Malawi regions boundaries}
\usage{
get_regions(layer = get_boundaries_name(adm = 1))
}
\arguments{
\item{layer}{Name of layer for regions boundaries}
}
\value{
An sf object for the regions borders of Malawi
}
\description{
Retrieve Malawi regions boundaries
}
\examples{
get_regions()

}
