% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_boundaries.R
\name{get_ta_areas}
\alias{get_ta_areas}
\title{Retrieve Malawi traditional authority areas boundaries}
\usage{
get_ta_areas(id)
}
\arguments{
\item{id}{Character vector for unique identifier of administrative
boundaries data for Malawi from Humanitarian Data Exchange}
}
\value{
An sf object for the traditional authority areas borders of Malawi
}
\description{
Retrieve Malawi traditional authority areas boundaries
}
\examples{
get_ta_areas(id = "cod-ab-mwi")

}
