% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_boundaries.R
\name{get_country}
\alias{get_country}
\title{Retrieve Malawi country boundaries}
\usage{
get_country(id)
}
\arguments{
\item{id}{Character vector for unique identifier of administrative
boundaries data for Malawi from Humanitarian Data Exchange}
}
\value{
An sf object for the country borders of Malawi
}
\description{
Retrieve Malawi country boundaries
}
\examples{
get_country(id = "cod-ab-mwi")

}
