% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_boundaries.R
\name{get_regions}
\alias{get_regions}
\title{Retrieve Malawi regions boundaries}
\usage{
get_regions(id)
}
\arguments{
\item{id}{Character vector for unique identifier of administrative
boundaries data for Malawi from Humanitarian Data Exchange}
}
\value{
An sf object for the regions borders of Malawi
}
\description{
Retrieve Malawi regions boundaries
}
\examples{
get_regions(id = "cod-ab-mwi")

}
