% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_boundaries.R
\name{get_ta_areas}
\alias{get_ta_areas}
\title{Retrieve Malawi traditional authority areas boundaries}
\usage{
get_ta_areas(layer = get_boundaries_name(adm = 3))
}
\arguments{
\item{layer}{Name of layer for traditional authority areas boundaries}
}
\value{
An sf object for the traditional authority areas borders of Malawi
}
\description{
Retrieve Malawi traditional authority areas boundaries
}
\examples{
get_ta_areas()

}
